Extract mesh serialization module
Move serialized-coordinate identity out of grid_tesselate: edge signatures,
coordinate/vertex/polygon/quad normalization, the serialized vertex cache,
triangle edge counts, and the collapse checks.

These answer final-output questions rather than building geometry, and every
nudge module needs them, so they belong below the layer that constructs cells.

Emitted meshes are unchanged: the four smoke configs stay byte identical.
